refactor authentication module with dedicated client/server export in lib folder
