+++ src/lib/auth/client.ts
@@ -1,7 +1,13 @@
 import { twoFactorClient } from "better-auth/plugins";
 import { createAuthClient } from "better-auth/react";
 
-export const authClient = createAuthClient({
+export const {
+  getSession,
+  signUp,
+  signOut,
+  forgetPassword: forgotPassword,
+  resetPassword,
+} = createAuthClient({
   baseURL: process.env.NEXT_PUBLIC_BASE_URL,
-  plugins: [twoFactorClient()]
+  plugins: [twoFactorClient()],
 });
